Refactors user management API and UI components
Updates the user fetching logic to retrieve data from the admin endpoint.

Replaces the user table component with a custom data table for better flexibility and functionality.

Removes unused user-related functions and enhances code organization by introducing a new users column definition file.

Improves usability by adding filtering capabilities to the data table.

Cleans up unnecessary console logs and optimizes data fetching in the custom hook.
