--- frontend/src/features/Admin/components/TableOrder.tsx
@@ -1,299 +1,0 @@
-"use client";
-
-import {
-  flexRender,
-  getCoreRowModel,
-  getFilteredRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  useReactTable,
-} from "@tanstack/react-table";
-import type { ColumnDef, ColumnFiltersState, SortingState, VisibilityState } from "@tanstack/react-table";
-import { ArrowUpDown, ChevronDown, Funnel, MoreHorizontal, Search } from "lucide-react";
-import * as React from "react";
-
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { Checkbox } from "@/components/ui/checkbox";
-import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { Input } from "@/components/ui/input";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
-
-export type Order = {
-  id: string;
-  customer: string;
-  date: string;
-  amount: string;
-  status?: "ĐÃ GIAO" | "CHỜ XỬ LÝ" | "HỦY" | "ĐANG XỬ LÝ";
-};
-
-const data: Order[] = [
-  { id: "INV001", customer: "Nguyễn Tuấn Cặc", date: "11-8-2003", amount: "250", status: "ĐÃ GIAO" },
-  { id: "INV002", customer: "Trần Thị B", date: "2024-05-02", amount: "180", status: "ĐANG XỬ LÝ" },
-  { id: "INV003", customer: "Lê Thị C", date: "2024-05-03", amount: "350", status: "HỦY" },
-  { id: "INV004", customer: "Phạm Văn D", date: "2024-05-04", amount: "420", status: "ĐÃ GIAO" },
-  { id: "INV005", customer: "Phạm Văn D", date: "2024-05-04", amount: "420", status: "ĐÃ GIAO" },
-  { id: "INV006", customer: "Phạm Văn D", date: "2024-05-04", amount: "420", status: "ĐÃ GIAO" },
-  { id: "INV007", customer: "Phạm Văn D", date: "2024-05-04", amount: "420", status: "CHỜ XỬ LÝ" },
-];
-
-export const columns: ColumnDef<Order>[] = [
-  {
-    id: "select",
-    header: ({ table }) => (
-      <Checkbox
-        checked={table.getIsAllPageRowsSelected() || (table.getIsSomePageRowsSelected() && "indeterminate")}
-        onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
-        aria-label="Select all"
-      />
-    ),
-    cell: ({ row }) => (
-      <Checkbox
-        checked={row.getIsSelected()}
-        onCheckedChange={(value) => row.toggleSelected(!!value)}
-        aria-label="Select row"
-      />
-    ),
-    enableSorting: false,
-    enableHiding: false,
-  },
-  {
-    accessorKey: "id",
-    header: "Mã đơn hàng",
-  },
-  {
-    accessorKey: "customer",
-    header: ({ column }) => (
-      <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}>
-        <span>Khách hàng</span> <ArrowUpDown className="ml-2 h-4 w-4" />
-      </Button>
-    ),
-    cell: ({ row }) => {
-      const customer = String(row.getValue("customer"));
-      return <div className="ml-3 py-4 font-medium">{customer}</div>;
-    },
-  },
-  {
-    accessorKey: "date",
-    header: ({ column }) => (
-      <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}>
-        Ngày <ArrowUpDown className="ml-2 h-4 w-4" />
-      </Button>
-    ),
-    cell: ({ row }) => {
-      const date = String(row.getValue("date"));
-      return <div className="ml-3 py-4 font-medium">{date}</div>;
-    },
-  },
-  // {
-  //   accessorKey: "amount",
-  //   header: ({ column }) => (
-  //     <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}>
-  //       Tổng tiền <ArrowUpDown className="ml-2 h-4 w-4" />
-  //     </Button>
-  //   ),
-  //   // cell: ({ row }) => {
-  //   //   const formatted = new Intl.NumberFormat("vi-VN", {
-  //   //     style: "currency",
-  //   //     currency: "VND",
-  //   //   }).format(row.getValue("amount"));
-  //   //   return <div className="text-right">{formatted}</div>;
-  //   // },
-  // },
-  {
-    accessorKey: "amount",
-    header: ({ column }) => (
-      <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}>
-        Amount <ArrowUpDown className="ml-2 h-4 w-4" />
-      </Button>
-    ),
-    cell: ({ row }) => {
-      const amount = parseFloat(row.getValue("amount"));
-
-      // Format the amount as a dollar amount
-      const formatted = new Intl.NumberFormat("en-US", {
-        style: "currency",
-        currency: "USD",
-      }).format(amount);
-
-      return <div className="ml-3 py-4 font-medium">{formatted}</div>;
-    },
-  },
-
-  {
-    accessorKey: "status",
-    header: ({ column }) => (
-      <Button variant="ghost" onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}>
-        Trạng thái <ArrowUpDown className="ml-2 h-4 w-4" />
-      </Button>
-    ),
-    cell: ({ row }) => {
-      const status = String(row.getValue("status"));
-      return (
-        <div className="px-2 py-4 font-medium">
-          <Badge
-            variant={
-              status === "ĐÃ GIAO"
-                ? "complete"
-                : status === "ĐANG XỬ LÝ"
-                  ? "pending"
-                  : status === "CHỜ XỬ LÝ"
-                    ? "waiting"
-                    : status === "HỦY"
-                      ? "danger"
-                      : "default"
-            }
-            className="w-2/5 px-3 text-sm font-medium capitalize"
-          >
-            {status}
-          </Badge>
-        </div>
-      );
-    },
-  },
-  {
-    id: "actions",
-    enableHiding: false,
-    enableSorting: false,
-
-    cell: () => {
-      return (
-        <div className="text-right font-medium">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="ghost" className="h-8 w-8 p-0">
-                <span className="sr-only">Open menu</span>
-                <MoreHorizontal />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end">
-              {/* <DropdownMenuLabel>Actions</DropdownMenuLabel>
-              <DropdownMenuItem onClick={() => navigator.clipboard.writeText(order.id)}>Copy Order ID</DropdownMenuItem>
-              <DropdownMenuSeparator /> */}
-              <DropdownMenuItem>Xem chi tiết</DropdownMenuItem>
-              <DropdownMenuItem>Hạn chế</DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-        </div>
-      );
-    },
-  },
-];
-
-export function TableOrder() {
-  const [sorting, setSorting] = React.useState<SortingState>([]);
-  const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>([]);
-  const [columnVisibility, setColumnVisibility] = React.useState<VisibilityState>({});
-  const [rowSelection, setRowSelection] = React.useState({});
-
-  const table = useReactTable({
-    data,
-    columns,
-    state: {
-      sorting,
-      columnFilters,
-      columnVisibility,
-      rowSelection,
-    },
-    onSortingChange: setSorting,
-    onColumnFiltersChange: setColumnFilters,
-    onColumnVisibilityChange: setColumnVisibility,
-    onRowSelectionChange: setRowSelection,
-    getCoreRowModel: getCoreRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
-  });
-
-  return (
-    <div className="w-full">
-      <div className="bg-card text-card-foreground flex items-center gap-2 rounded-lg border p-4">
-        <div className="flex-grow">
-          <div className="relative flex items-center">
-            <Search className="text-muted-foreground absolute left-3 h-4 w-4" />
-
-            <Input
-              placeholder="Tìm sản phẩm..."
-              value={(table.getColumn("customer")?.getFilterValue() as string) ?? ""}
-              onChange={(event) => table.getColumn("customer")?.setFilterValue(event.target.value)}
-              className="p-6 pl-10" // Giữ padding gốc của bạn và thêm pl-10
-            />
-          </div>
-        </div>
-        <div>
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="outline" className="ml-auto p-6">
-                <span className="text-sm">
-                  <Funnel className="" />
-                </span>
-                <span className="text-sm">Bộ Lộc</span>
-              </Button>
-            </DropdownMenuTrigger>
-          </DropdownMenu>
-        </div>
-      </div>
-
-      <div className="mt-5 rounded-md border">
-        <Table>
-          <TableHeader>
-            {table.getHeaderGroups().map((headerGroup) => (
-              <TableRow key={headerGroup.id} className="">
-                {headerGroup.headers.map((header) => (
-                  <TableHead key={header.id}>
-                    {header.isPlaceholder ? null : flexRender(header.column.columnDef.header, header.getContext())}
-                  </TableHead>
-                ))}
-              </TableRow>
-            ))}
-          </TableHeader>
-          <TableBody>
-            {table.getRowModel().rows.length ? (
-              table.getRowModel().rows.map((row) => (
-                <TableRow key={row.id} data-state={row.getIsSelected() && "selected"}>
-                  {row.getVisibleCells().map((cell) => (
-                    <TableCell key={cell.id} className="px-2">
-                      {flexRender(cell.column.columnDef.cell, cell.getContext())}{" "}
-                    </TableCell>
-                  ))}
-                </TableRow>
-              ))
-            ) : (
-              <TableRow>
-                <TableCell colSpan={columns.length} className="h-24 text-center">
-                  Không có kết quả.
-                </TableCell>
-              </TableRow>
-            )}
-          </TableBody>
-        </Table>
-      </div>
-
-      <div className="flex items-center justify-end space-x-2 py-4">
-        <div className="text-muted-foreground flex-1 text-sm">
-          {table.getFilteredSelectedRowModel().rows.length} / {table.getFilteredRowModel().rows.length} đơn được chọn.
-        </div>
-        <div className="space-x-2">
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => table.previousPage()}
-            disabled={!table.getCanPreviousPage()}
-          >
-            Trước
-          </Button>
-          <Button variant="outline" size="sm" onClick={() => table.nextPage()} disabled={!table.getCanNextPage()}>
-            Tiếp
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-}
-export default TableOrder;
